use `build.py`
